backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
anthropic==0.40.0
openai>=1.0.0
faster-whisper==1.0.3
requests==2.32.5
python-docx==1.1.2
pydantic==2.9.2
pydantic-settings==2.5.2
